services/user-service/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
pydantic[email]
mysqlclient==2.2.7
passlib[bcrypt]
bcrypt<4.1.0
python-jose==3.4.0
python-multipart>=0.0.22
pymysql==1.1.1
cryptography>=46.0.5
pika==1.3.2
httpx
pytest
